+++ src/main.go
@@ -9,7 +9,7 @@ import (
 
 // nolint: gochecknoglobals
 var (
-	version = "0.3.0"
+	version = "0.3.1"
 	commit  = ""
 	date    = ""
 	builtBy = ""